==> jupyterNotebooks/1 - Exploração inicial dos datasets utilizados.ipynb <==
# PUC MINAS

## Pós Graduação em Ciência de Dados e Big Data (2019-2020)

### Trabalho de Conclusão de Curso

### Aluna: Maria da Glória dos Santos Batista de Oliveira




Este notebook é relacionado ao meu trabalho de conclusão de curso em Ciência de Dados e Big Data, o título é Análises de Séries Temporais Relacionadas a Covid19 no Brasil, neste trabalho os datasets foran baixados no site [Kaggle](https://www.kaggle.com/), especificamente os arquivos a seguir:

- brazil_covid19_macro.csv
- brazil_covid19_cities.csv
- brazil_cities_coordinates.csv

Para a realização dos downloads dos datasets, clique [aqui](https://www.kaggle.com/unanimad/corona-virus-brazil/data?select=brazil_covid19_cities.csv).

__Leitura dos dados e importação das bibliotecas utilizadas__

Para iniciar o trabalho, primeiramente é necessário fazer download dos datasets informados previamente. Assim o documento poderá ser lido sem que qualquer problema ocorra.

No código a seguir, as importações das bibliotecas utilizadas e a definição dos métodos para facilitar a visualização dos dados e exibição de gráficos de value counts no Pandas.

# Imports das bibliotecas utilizadas

import pandas as pd

import warnings
warnings.filterwarnings('ignore')

Vamos realizar as importações dos datasets utilizados:

# Import do primeiro dataset utilizado

arquivo = '../datasets_originais/datasets_549702_1421038_brazil_covid19_macro.csv'
df = pd.read_csv(arquivo, sep=',', parse_dates=['date'])

# Import do segundo dataset utilizado

cidades = '../datasets_originais/brazil_covid19_cities.csv.zip'
df_cidades = pd.read_csv(cidades, sep=',',parse_dates=['date'], compression='zip')

# Import do terceiro dataset utilizado

coord_cidades = '../datasets_originais/datasets_549702_1423724_brazil_cities_coordinates.csv'
df_coord_cidades = pd.read_csv(coord_cidades, sep=',')

Visualizando o primeiro dataset criado:

# Visualizando as 5 primeiras linhas do arquivo

df.head(5)

#### Dicionário de dados:

- Date: Data
- Country: País
- Week: Número da Semana
- Cases: Número de casos confirmados
- Deaths: Número de Mortes
- Recovered: Número de pessoas recuperadas
- Monitoring: Número de pessoas monitoradas

Visualizando o segundo dataset criado:

#Visualizando o dataset criado

df_cidades.head(5)

#### Dicionário de dados:

- Date: Data
- State: Estado
- Name: Cidade
- Code: Código
- Casos: Número de pessoas infectadas
- Deaths: Número de Mortes



Visualizando o terceiro dataset criado:

# Visualizando 

df_coord_cidades.head(5)

Dicionário de dados:

- state_code: Código do estado
- city_code: Código da cidade
- city_name: Nome da cidade
- lat: Latitude
- long: Longitude
- capital: Indicador de capital

### Salvando os novos datasets:

# Salvando o dataset
df.to_csv('../datasets_originais/df.csv')

# Salvando o dataset
df_cidades.to_csv('../datasets_originais/df_cidades.csv.gz', compression='gzip')

# Salvando o dataset
df_coord_cidades.to_csv('../datasets_originais/df_coord_cidades.csv')
