==> jupyterNotebooks/2 - Processamento e Tratamento de Dados.ipynb <==
# PUC MINAS

## Pós Graduação em Ciência de Dados e Big Data (2019-2020)

### Trabalho de Conclusão de Curso

### Aluna: Maria da Glória dos Santos Batista de Oliveira




__Leitura dos dados e importação das bibliotecas utilizadas__

Para iniciar o trabalho, primeiramente é necessário fazer download dos datasets informados previamente. Assim o documento poderá ser lido sem que qualquer problema ocorra.

No código a seguir, as importações das bibliotecas utilizadas e a definição dos métodos para facilitar a visualização dos dados e exibição de gráficos de value counts no Pandas.

# Imports das bibliotecas utilizadas

import pandas as pd
import numpy as np

# Não exibir as mensagens de warnings
import warnings
warnings.filterwarnings('ignore')

Vamos realizar as importações dos datasets utilizados:

# Import do primeiro dataset utilizado

arquivo = '../datasets_originais/df.csv'

# filtrando as colunas que serão utilizadas:
lista_colunas = ["date", "cases", "deaths", "recovered", "monitoring"]

df = pd.read_csv(arquivo, sep=',', parse_dates=['date'], usecols=lista_colunas)

# Import do segundo dataset utilizado

cidades = '../datasets_originais/df_cidades.csv.gz'

# filtrando as colunas que serão utilizadas:
lista_colunas = ["date", "state", "name", "cases", "deaths"]

df_cidades = pd.read_csv(cidades, sep=',',parse_dates=['date'], compression='gzip', usecols=lista_colunas)

# Import do terceiro dataset utilizado

coord_cidades = '../datasets_originais/df_coord_cidades.csv'

# filtrando as colunas que serão utilizadas:
lista_colunas = ["city_name", "lat", "long"]

df_coord_cidades = pd.read_csv(coord_cidades, sep=',', usecols=lista_colunas)

#### Vamos visualizar os datasets importados:

# Visualizando as 5 primeiras linhas:

df.head(5)

# Visualizando as 5 primeiras linhas:

df_cidades.head(5)

# Visualizando as 5 primeiras linhas:
df_coord_cidades.head(5)

#### Vamos fazer a verificação se os dataframes possuem alguma coluna com valor nulo:

# Verificando valores nulos

df.isna().any()

Como podemos observar, existem duas colunas que possuem valores nulos: 
- recovered 
- monitoring

Verificando se no dataframe existem valores duplicados:

# Verificação de duplicidade

ind_duplicado = df.duplicated().any()

Agora vamos ver se existem valores duplicados:

# Visualizando se existem duplicidades

print(ind_duplicado)

Como podemos observar não há registro duplicado no dataset.

# Verificando valores nulos no dataframe sobre cidades

df_cidades.isna().any()

Como podemos observar, não existem colunas que possuam valores nulos.

Verificando se o dataframe existem valores duplicados:

# Verificação de duplicidade

ind_duplicado = df_cidades.duplicated().any()

Agora vamos ver se existem valores duplicados:

# Visualizando se existem duplicidades

print(ind_duplicado)

Como podemos observar não existem registros duplicados.

Agora iremos verificar se existem valores nulos no dataset:

# Verificando valores nulos

df_coord_cidades.isna().any()

Como podemos ver não existem registros nulos no dataframe.

Verificando registros duplicados no dataframe:

# Verificação de duplicidade
ind_duplicado = df_coord_cidades.duplicated().any()

Visualizando se existem registros duplicados:

# Visualizando se existem duplicidades

print(ind_duplicado)

Como podemos observar não há registro duplicado no dataset.

#### Vamos realizar o tratamento do dataframe que possui valores nulos:

# Resolvendo o problema de valores nulos no dataframe: preenchendo valores nulos com zero

df_corona=df.fillna(0)

___Agora que os dados nulos foram tratados dentro do dataframe, será necessário realizar a verificação de como o pandas identificou os dados no momento da leitura do dataset, será utilizado o comando info() para verificar essas informações.___

# Visualizando a informação sobre os datatypes

df_corona.info()

Como podemos observar, será necessário mudar o datatype de duas colunas: 
- recovered 
- monitoring

Vamos realizar o tratamento para que todas as colunas numéricas possuam o mesmo datatype (int).

# Mudando o datatype de Float para Int e criando um novo dataframe

df_corona=df_corona.astype({"recovered": int, "monitoring": int})

Agora vamos verificar se os datatypes foram realmente alterados nesse novo dataframe.

# Visualizando os dados

df_corona.info()

Realizando a verificação se haverá necessidade de alterar os datatypes do segundo dataset:

# Mostra a informação do dataframe

df_cidades.info()

Não havera a necessidade de alterar nenhum datatype do dataframe.

Realizando a verificação se haverá necessidade de alterar os datatypes do terceiro dataset:

# Visualizando a informação sobre os datatypes
df_coord_cidades.info()

Como podemos verificar não haverá a necessidade de nenhuma alteração nos datatypes.

#### Incluindo um novo indicador no dataframe

A partir dos dados importados do dataset, vamos criar uma nova coluna para análise no dataframe:

Casos ativos = Casos – Mortes – Recuperados

Neste contexto, vamos analisar uma pessoa que contraiu o Coronavírus, ela não morreu porém ainda não se recuperou da doença.

# Criando uma nova coluna

df_corona['active'] = df_corona['cases'] - df_corona['deaths'] - df_corona['recovered']

Agora vamos visualizar o dataframe modificado:

# Visualizando os dados

df_corona.info()

Agora será necessário verificar os dados temporais contidos nos dataframes.

# Verifica os dados temporais do dataframe

df_corona.date.describe()

# Verifica os dados temporais do dataframe

df_cidades.date.describe()

Com a informação obtida dos dataframes, é possível observar a data de início e data de término contidas nos arquivos.

### Agrupando os dados

Os dados serão mostrados de maneira que o número casos, mortes, monitorados, recuperados e ativos sejam agrupados por data.

# Agrupando os dados

df_agrupado = df_corona.groupby(['date'])['cases','deaths','monitoring', 'recovered', 'active'].sum().reset_index()

O dataframe será ordenado para descobrir em qual data ocorreu o maior número de casos ativos.

# Ordenando os registros

df_agrupado.sort_values(by='active', ascending=False)

Agrupando a quantidade de casos recuperados, mortes e ativos por data.

# Agrupando

df_agrupado = df_agrupado.groupby('date')['cases','deaths','monitoring','recovered','active'].sum().reset_index();

Será utilizada a função melt do pandas para remodelar o dataframe com variável e valor, para ter a quantidade de casos recuperados, mortos e ativos.

# Função melt do pandas para remodelar o dataframe com variável e valor, para ter a quantidade de casos recuperados, mortos e ativos.

df_casos_agrupados = df_agrupado.melt(id_vars="date", value_vars=['recovered', 'deaths', 'monitoring','active'],
                 var_name='case', value_name='count')

Visualizando as 5 primeiras linhas do dataframe "df_casos_agrupados":

# Visualizando as 5 primeiras linhas

df_casos_agrupados.head(5)

Visualizando 5 últimas linhas do dataframe "df_casos_agrupados":

# Visualizando as 5 últimas linhas

df_casos_agrupados.tail(5)

Agrupando os dados da geolocalização:

# Visualizando as 5 primeiras linhas

df_cidades.head(5)

Agrupando a quantidade de casos, mortes por data, estado e nome.

# Agrupando pela quantidade casos, mortes por data, estado e nome

df_cidades_agrupado = df_cidades.groupby(['date','state','name'])['cases','deaths'].sum().reset_index()

Criando um novo dataset a partir dos dasets df_cidades_agrupado e df_coord_cidades:

# Usando Merge que contenha a cidades que tenham o mesmo nome

df_geo_cidades_temp = pd.merge(left=df_cidades_agrupado, right=df_coord_cidades, left_on='name', right_on='city_name')

Agora criaremos um novo dataset sobre geolocalização com informações que serão relevantes para a análise.

# Criando um novo dataset que contenha somente as informações que preciso

df_geo_cidades = df_geo_cidades_temp[['date', 'state','name','cases','deaths','lat','long']].copy()

Visualizando o novo dataset:

# Visualizando o novo dataset

df_geo_cidades.head(5)

Ordenando os dados pela quantidade de casos.

# Ordenando pela quantidade de casos

df_geo_cidades.sort_values(by='cases', ascending=False)

Agora retornando para o dataset "df_cidades_agrupado", vamos agrupar as quantidades de casos e mortos por data, estado e nome.

# Agrupando o dataframe

df_cidades_agrupado = df_cidades_agrupado.groupby(['date','state','name'])['cases', 'deaths'].sum().reset_index();

Agora vamos visualizar o dataframe criado:

# Visualizando os dados contidos no dataframe

df_cidades_agrupado.head(5)


### Salvando os datasets tratados:

# Salvando o dataset
df_corona.to_csv('../datasets_tratados/df_corona.csv')

# Salvando o dataset
df_agrupado.to_csv('../datasets_tratados/df_agrupado.csv')

# Salvando o dataset
df_casos_agrupados.to_csv('../datasets_tratados/df_casos_agrupados.csv')

# Salvando o dataset
df_cidades_agrupado.to_csv('../datasets_tratados/df_cidades_agrupado.csv.gz', compression='gzip')

# Salvando o dataset
df_geo_cidades.to_csv('../datasets_tratados/df_geo_cidades.csv.gz', compression='gzip')
