==> jupyterNotebooks/3 - Processamento e Tratamento de Dados.ipynb <==
# PUC MINAS

## Pós Graduação em Ciência de Dados e Big Data (2019-2020)

### Trabalho de Conclusão de Curso

### Aluna: Maria da Glória dos Santos Batista de Oliveira




__Leitura dos dados e importação das bibliotecas utilizadas__

Para iniciar o trabalho, primeiramente é necessário fazer download dos datasets informados previamente. Assim o documento poderá ser lido sem que qualquer problema ocorra.

No código a seguir, as importações das bibliotecas utilizadas e a definição dos métodos para facilitar a visualização dos dados e exibição de gráficos de value counts no Pandas.

# Imports das bibliotecas utilizadas

import pandas as pd
import numpy as np

# Não exibir as mensagens de warnings
import warnings
warnings.filterwarnings('ignore')

Vamos realizar a importação do dataset utilizado:

# Import do primeiro dataset utilizado
arquivo = '../datasets_tratados/df_agrupado.csv'

# Filtrando as colunas que serão utilizadas:
lista_colunas = ["date", "cases", "deaths", "monitoring", "recovered", "active"]

df_agrupado = pd.read_csv(arquivo, sep=',', parse_dates=['date'], usecols=lista_colunas)

#### Visualizando o dataset importado:

# Visualizando o dataset:

df_agrupado.head(5)

#### - Quantas pessoas foram infectadas com esta doença?

Para uma análise melhor dos dados, vamos gerar um dataframe exclusivo para responder esta pergunta.

# Gerando dataframe para análise de pessoas doentes

df_qtd_ativos = df_agrupado.melt(id_vars="date", value_vars=['active'],
                 var_name='case', value_name='count')

#### - Quantas pessoas estão sendo monitoradas com suspeita dessa doença?

Para uma realização de uma melhor análise dos dados, vamos gerar um dataframe exclusivo para responder esta pergunta.

# Gerando dataframe para análise de pessoas monitoradas

df_qtd_monitorados = df_agrupado.melt(id_vars="date", value_vars=['monitoring'],
                 var_name='case', value_name='count')

#### - Quantas pessoas morreram com esta doença?

Para um melhor entendimento dos dados, vamos gerar um dataframe exclusivo para responder esta pergunta.

# Gerando dataframe para análise de pessoas mortas

df_qtd_mortos = df_agrupado.melt(id_vars="date", value_vars=['deaths'],
                 var_name='case', value_name='count')


#### - Quantas pessoas se recuperaram com esta doença?

Para uma melhor compreensão dos dados, vamos gerar um dataframe exclusivo para responder esta pergunta.

# Gerando dataframe para análise de pessoas recuperadas

df_qtd_recuperados = df_agrupado.melt(id_vars="date", value_vars=['recovered'],
                 var_name='case', value_name='count')

### -  Geração de datasets sobre o Corona Vírus para os seguintes cenários:

#### - Quantidade pessoas que terão essa doença

# Criando um novo dataframe para analise a partir do dataframe df_ativos

df_ativos = df_qtd_ativos[['date', 'count']].copy()

#### - Quantidade de pessoas que se recuperarão desta doença


# Montando um dataframe para análise dos dados

df_recuperados = df_qtd_recuperados[['date', 'count']].copy()

#### - Quantidade de pessoas que morrerão com esta doença

# Montando um dataframe para análise dos dados

df_mortos = df_qtd_mortos[['date', 'count']].copy()

### Salvando os datasets tratados:

# Salvando o dataset

df_qtd_ativos.to_csv('../datasets_tratados/df_qtd_ativos.csv')

# Salvando o dataset

df_qtd_monitorados.to_csv('../datasets_tratados/df_qtd_monitorados.csv')

# Salvando o dataset

df_qtd_mortos.to_csv('../datasets_tratados/df_qtd_mortos.csv')

# Salvando o dataset

df_qtd_recuperados.to_csv('../datasets_tratados/df_qtd_recuperados.csv')

# Salvando o dataset

df_ativos.to_csv('../datasets_tratados/df_ativos.csv')

# Salvando o dataset

df_recuperados.to_csv('../datasets_tratados/df_recuperados.csv')

# Salvando o dataset

df_mortos.to_csv('../datasets_tratados/df_mortos.csv')
