==> jupyterNotebooks/4 - Análise de dados - Parte 2.ipynb <==
# PUC MINAS

## Pós Graduação em Ciência de Dados e Big Data (2019-2020)

### Trabalho de Conclusão de Curso

### Aluna: Maria da Glória dos Santos Batista de Oliveira




__Leitura dos dados e importação das bibliotecas utilizadas__

Para iniciar o trabalho, primeiramente é necessário fazer download dos datasets informados previamente. Assim o documento poderá ser lido sem que qualquer problema ocorra.

No código a seguir, as importações das bibliotecas utilizadas e a definição dos métodos para facilitar a visualização dos dados e exibição de gráficos de value counts no Pandas.

# Imports das bibliotecas utilizadas

import pandas as pd
import numpy as np

# Visualização
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
import plotly.express as px
import plotly.graph_objects as go

# Manipulando o tamanho padrão do plot
plt.rcParams['figure.figsize'] = 8, 5

# Não exibir as mensagens de warnings
import warnings
warnings.filterwarnings('ignore')

Vamos realizar as importações dos datasets utilizados:

# Import do dataset utilizado

cidades_agrupado = '../datasets_tratados/df_cidades_agrupado.csv.gz'

# filtrando as colunas que serão utilizadas:
lista_colunas = ["date","state", "name", "cases","deaths"]

df_cidades_agrupado = pd.read_csv(cidades_agrupado, sep=',', parse_dates=['date'], compression='gzip', usecols=lista_colunas)

#### Visualizando os datasets importados:

# Visualizando o dataset:

df_cidades_agrupado.head(5)

#### 6 - Qual a região mais afetada por esta doença?

Criando um dataframe para posteriormente um gráfico ser criado a partir dele.

# Ordena por data, casos e mortes

df_cidades_agrupado.sort_values(by=['date','cases','deaths'],
        ascending=[False, False, False]).head()


Visualizando o gráfico por estado relacionado à esta doença:

# Criando o gráfico

fig = px.line(df_cidades_agrupado,
              x="date",
              y="cases",
              color='state',
              height=600,
              title='Casos Confirmados',
              color_discrete_sequence = px.colors.qualitative.Dark2 )
fig.show()


fig = px.line(df_cidades_agrupado,
              x="date",
              y="deaths",
              color='state',
              height=600,
              title='Mortes Confirmadas',
              color_discrete_sequence = px.colors.qualitative.Dark2)
fig.show()
